提交人：王刚 提交信息：用户审核，checkUser 提交时间：4-26 18：04

+++ src/main/java/com/aaa/project/system/userAccount/service/UserAccountServiceImpl.java
@@ -47,7 +47,7 @@ public class UserAccountServiceImpl implements IUserAccountService {
      * @return 用户信息
      */
     @Override
-    public UserAccount selectUserAccountByUserId(Integer userId) {
+    public UserAccount selectUserAccountByUserId(Long userId) {
         return userAccountMapper.selectUserAccountByUserId(userId);
     }
